[Silver II] Title: 랜선 자르기, Time: 184 ms, Memory: 17420 KB -BaekjoonHub

+++ 백준/Silver/1654. 랜선 자르기/랜선 자르기.java
@@ -52,8 +52,8 @@ public class Main {
 		return candidates.get(0);
 	}
 	
-	private static long getLineCnt(int[] lines, long cutLength) {
-		long lineCnt = 0;
+	private static int getLineCnt(int[] lines, long cutLength) {
+		int lineCnt = 0;
 		
 		for (int i = 0; i < lines.length; i++) {
 			lineCnt += (long) lines[i] / cutLength;
